requirements.txt
Flask
Flask-SQLAlchemy
flask-apispec
marshmallow